+++ database/nostr-database/CHANGELOG.md
@@ -23,6 +23,20 @@
 
 -->
 
+## Unreleased
+
+### Breaking changes
+
+- Remove `Backend::is_persistent` method (https://github.com/rust-nostr/nostr/pull/1138)
+
+### Changed
+
+- Extend `Backend` variants (https://github.com/rust-nostr/nostr/pull/1138)
+
+### Added
+
+- Add `Backend::custom` constructor (https://github.com/rust-nostr/nostr/pull/1138)
+
 ## v0.44.0 - 2025/11/06
 
 No notable changes in this release.